♻️ Sign_up should be its own feature
