--- src/features/greeting/mod.rs
@@ -1,1 +1,0 @@
-pub mod greeting_controller;
