+++ src/features/mod.rs
@@ -3,3 +3,4 @@ pub mod equipment;
 pub mod user;
 pub mod auth;
 pub mod protected;
+pub mod sign_up;
