+++ src/features/sign_up/controllers/mod.rs
@@ -1,0 +1,1 @@
+pub mod sign_up_controller;
