+++ src/features/sign_up/mod.rs
@@ -1,0 +1,1 @@
+pub mod controllers;
